fix: fixes for outdated stuff

+++ bootstrap/archlinux.sh
@@ -11,6 +11,11 @@ rm /opt/microsoft/powershell/7 -rf
 mkdir -p /opt/microsoft/powershell/7
 tar zxf /tmp/powershell.tar.gz -C /opt/microsoft/powershell/7
 chmod +x /opt/microsoft/powershell/7/pwsh
-rm /usr/bin/pwsh
+
+if [ -f /usr/bin/pwsh ] ; then
+    rm /usr/bin/pwsh
+fi
+
+
 ln -s /opt/microsoft/powershell/7/pwsh /usr/bin/pwsh
